requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
Django==2.2.1
django-countries==5.3.3
django-taggit==1.1.0
Pillow==2.2.1
psycopg2==2.8.2
pycparser==2.19
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
